Clean up client by moving to transport
