+++ src/client.rs
@@ -4,80 +4,10 @@
 //   file, You can obtain one at http://mozilla.org/MPL/2.0/.
 //
 
-use futures::AsyncRead;
-use futures::AsyncWrite;
-use tokio::io::DuplexStream;
-use tokio::net::TcpStream;
-use tokio_util::compat::Compat as TokioCompat;
-use tokio_util::compat::TokioAsyncReadCompatExt;
-
 use crate::client_identifier::ClientIdentifier;
 use crate::keep_alive::KeepAlive;
-
-pub(crate) enum MqttConnection {
-    Tokio(TokioCompat<tokio::net::TcpStream>),
-    Duplex(TokioCompat<tokio::io::DuplexStream>),
-}
-
-impl AsyncRead for MqttConnection {
-    fn poll_read(
-        mut self: std::pin::Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-        buf: &mut [u8],
-    ) -> std::task::Poll<std::io::Result<usize>> {
-        match &mut *self {
-            MqttConnection::Tokio(t) => std::pin::pin!(t).poll_read(cx, buf),
-            MqttConnection::Duplex(d) => std::pin::pin!(d).poll_read(cx, buf),
-        }
-    }
-}
-
-impl AsyncWrite for MqttConnection {
-    fn poll_write(
-        mut self: std::pin::Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-        buf: &[u8],
-    ) -> std::task::Poll<std::io::Result<usize>> {
-        match &mut *self {
-            MqttConnection::Tokio(t) => std::pin::pin!(t).poll_write(cx, buf),
-            MqttConnection::Duplex(d) => std::pin::pin!(d).poll_write(cx, buf),
-        }
-    }
-
-    fn poll_flush(
-        mut self: std::pin::Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-    ) -> std::task::Poll<std::io::Result<()>> {
-        match &mut *self {
-            MqttConnection::Tokio(t) => std::pin::pin!(t).poll_flush(cx),
-            MqttConnection::Duplex(d) => std::pin::pin!(d).poll_flush(cx),
-        }
-    }
-
-    fn poll_close(
-        mut self: std::pin::Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-    ) -> std::task::Poll<std::io::Result<()>> {
-        match &mut *self {
-            MqttConnection::Tokio(t) => std::pin::pin!(t).poll_close(cx),
-            MqttConnection::Duplex(d) => std::pin::pin!(d).poll_close(cx),
-        }
-    }
-}
-
-pub enum MqttConnectTransport {
-    TokioTcp(TcpStream),
-    TokioDuplex(DuplexStream),
-}
-
-impl From<MqttConnectTransport> for MqttConnection {
-    fn from(value: MqttConnectTransport) -> Self {
-        match value {
-            MqttConnectTransport::TokioTcp(t) => MqttConnection::Tokio(t.compat()),
-            MqttConnectTransport::TokioDuplex(d) => MqttConnection::Duplex(d.compat()),
-        }
-    }
-}
+use crate::transport::MqttConnectTransport;
+use crate::transport::MqttConnection;
 
 pub enum CleanStart {
     No,
@@ -135,12 +65,6 @@ pub struct MqttClientConnector {
 }
 
 impl MqttClientConnector {
-    pub fn properties_mut(&mut self) -> &mut ConnectProperties {
-        &mut self.properties
-    }
-}
-
-impl MqttClientConnector {
     pub fn new(
         transport: MqttConnectTransport,
         client_identifier: ClientIdentifier,
@@ -170,6 +94,10 @@ impl MqttClientConnector {
         };
 
         todo!()
+    }
+
+    pub fn properties_mut(&mut self) -> &mut ConnectProperties {
+        &mut self.properties
     }
 }
