+++ src/lib.rs
@@ -12,3 +12,4 @@ mod keep_alive;
 mod packet;
 mod properties;
 mod string;
+mod transport;
